A2/requirements.txt
Cython==0.25.2
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.0.0
Pygments==2.0.2
appnope==0.1.0
argparse==1.2.1
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
certifi==2015.11.20.1
cycler==0.10.0
decorator==4.0.6
future==0.16.0
gnureadline==6.3.3
h5py==2.7.0
ipykernel==4.2.2
ipython==4.0.1
ipython-genutils==0.1.0
ipywidgets==4.1.1
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.1.1
jupyter-console==4.0.3
jupyter-core==4.0.6
matplotlib==2.0.0
mistune==0.7.1
nbconvert==4.1.0
nbformat==4.0.1
nltk==3.2.2
notebook==4.0.6
numpy==1.12.1
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
pyparsing==2.0.7
python-dateutil==2.4.2
pytz==2015.7
pyzmq==15.1.0
qtconsole==4.1.1
scipy==0.19.0
simplegeneric==0.8.1
singledispatch==3.4.0.3

six==1.10.0
terminado==0.5
tornado==4.3
traitlets==4.0.0
